attempt fix loading in build
